requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
validators>=0.20.0